X11: Fix internal timer ID type to match public interface

+++ pugl/detail/x11.h
@@ -44,7 +44,7 @@ typedef struct {
 typedef struct {
 	XID       alarm;
 	PuglView* view;
-	uint64_t  id;
+	uintptr_t id;
 } PuglTimer;
 
 struct PuglWorldInternalsImpl {
